# 2.2.3/proxy/docs/swagger.yaml
basePath: /
definitions:
  controller.RequestAddressGeocode:
    properties:
      lat:
        example: "59.923013"
        type: string
      lng:
        example: "30.318105"
        type: string
    type: object
  controller.RequestAddressSearch:
    properties:
      query:
        example: Московский проспект 14
        type: string
    type: object
  controller.ResponseAddress:
    properties:
      addresses:
        items:
          $ref: '#/definitions/geoservice.Address'
        type: array
    type: object
  geoservice.Address:
    properties:
      city:
        example: Санкт-Петербург
        type: string
      house:
        example: "14"
        type: string
      lat:
        example: "59.923013"
        type: string
      lon:
        example: "30.318105"
        type: string
      street:
        example: Московский
        type: string
    type: object
  responder.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Поиск информации по адресу или координатам.
  title: ГеоAPI
  version: "1.0"
paths:
  /api/address/geocode:
    post:
      consumes:
      - application/json
      parameters:
      - description: Координаты. lat - Географическая широта. lng - Географическая
          долгота.
        in: body
        name: lat,lng
        required: true
        schema:
          $ref: '#/definitions/controller.RequestAddressGeocode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseAddress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responder.Response'
      summary: Обратное геокодирование (адрес по координатам)
      tags:
      - api
  /api/address/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Адрес
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/controller.RequestAddressSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseAddress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responder.Response'
      summary: Геокодирование (координаты по адресу)
      tags:
      - api
swagger: "2.0"